requirements.txt
streamlit
google-cloud-aiplatform
requests
google-cloud-discoveryengine